+++ Source/ChromeCast.Desktop.AudioStreamer/Application/Interfaces/IDevice.cs
@@ -34,7 +34,7 @@ namespace ChromeCast.Desktop.AudioStreamer.Application
         void Start();
         void OnReceiveMessage(CastMessage castMessage);
         DeviceControl GetDeviceControl();
-        ushort GetPort();
+        int GetPort();
         DiscoveredDevice GetDiscoveredDevice();
         void SendSilence();
         bool IsGroup();
